Extract data fetching actions
